# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CalendarSyncInput {
  /** @nullable */
  bookingCom?: string | null;
  /** @nullable */
  airbnb?: string | null;
  /** @nullable */
  makeMyTrip?: string | null;
}

export type CalendarEventSource = typeof CalendarEventSource[keyof typeof CalendarEventSource];


export const CalendarEventSource = {
  manual: 'manual',
  direct: 'direct',
  bookingCom: 'bookingCom',
  airbnb: 'airbnb',
  makeMyTrip: 'makeMyTrip',
} as const;

export interface CalendarEvent {
  id: string;
  source: CalendarEventSource;
  /** @nullable */
  title: string | null;
  /** @nullable */
  note: string | null;
  startDate: string;
  endDate: string;
  editable: boolean;
}

export type CalendarSourceStatusSource = typeof CalendarSourceStatusSource[keyof typeof CalendarSourceStatusSource];


export const CalendarSourceStatusSource = {
  bookingCom: 'bookingCom',
  airbnb: 'airbnb',
  makeMyTrip: 'makeMyTrip',
} as const;

export type CalendarSourceStatusStatus = typeof CalendarSourceStatusStatus[keyof typeof CalendarSourceStatusStatus];


export const CalendarSourceStatusStatus = {
  connected: 'connected',
  missing: 'missing',
  error: 'error',
} as const;

export interface CalendarSourceStatus {
  source: CalendarSourceStatusSource;
  label: string;
  status: CalendarSourceStatusStatus;
  eventCount: number;
  message: string;
  /** @nullable */
  lastSyncedAt: string | null;
}

export interface CalendarSyncResponse {
  syncedAt: string;
  totalEvents: number;
  events: CalendarEvent[];
  sources: CalendarSourceStatus[];
}

export interface CalendarFeedInfo {
  feedUrl: string;
  bookingCom: string;
  airbnb: string;
  makeMyTrip: string;
}

export interface CalendarBlockInput {
  startDate: string;
  endDate: string;
  /**
     * @maxLength 120
     * @nullable
     */
  title?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  note?: string | null;
}

export interface CalendarEventsResponse {
  events: CalendarEvent[];
}

export interface PublicCalendarBlock {
  startDate: string;
  endDate: string;
}

export interface PublicCalendarResponse {
  blocks: PublicCalendarBlock[];
}

export interface ErrorResponse {
  error: string;
}

export type GetCalendarFeedParams = {
token: string;
/**
 * Omit events imported from this source to prevent feedback loops.
 */
exclude?: GetCalendarFeedExclude;
};

export type GetCalendarFeedExclude = typeof GetCalendarFeedExclude[keyof typeof GetCalendarFeedExclude];


export const GetCalendarFeedExclude = {
  bookingCom: 'bookingCom',
  airbnb: 'airbnb',
  makeMyTrip: 'makeMyTrip',
} as const;