# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: calendar
    description: Central availability calendar operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /calendar/sync:
    post:
      operationId: syncCalendars
      tags: [calendar]
      summary: Sync OTA calendar feeds
      description: Fetches configured OTA iCal feeds and persists their blocked date snapshots.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarSyncInput"
      responses:
        "200":
          description: Calendar feeds synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarSyncResponse"
        "400":
          description: Invalid calendar feed configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin sign-in required
        "403":
          description: Account is not an authorised admin
        "503":
          description: Admin access is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calendar/events:
    get:
      operationId: listCalendarEvents
      tags: [calendar]
      summary: List calendar events for administrators
      responses:
        "200":
          description: Persisted calendar events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEventsResponse"
        "401":
          description: Admin sign-in required
        "403":
          description: Account is not an authorised admin
        "503":
          description: Admin access is not configured
  /calendar/public:
    get:
      operationId: getPublicCalendar
      tags: [calendar]
      summary: Read public blocked date ranges
      description: Returns only blocked ranges without sources, titles, or notes.
      responses:
        "200":
          description: Public blocked date ranges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCalendarResponse"
  /calendar/block:
    post:
      operationId: blockCalendarDates
      tags: [calendar]
      summary: Block dates manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarBlockInput"
      responses:
        "201":
          description: Calendar block created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
        "400":
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin sign-in required
        "403":
          description: Account is not an authorised admin
  /calendar/block/{id}:
    delete:
      operationId: unblockCalendarDates
      tags: [calendar]
      summary: Remove a manual calendar block
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Calendar block removed
        "400":
          description: Invalid event id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin sign-in required
        "403":
          description: Account is not an authorised admin
        "404":
          description: Event not found or not editable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /calendar/feed-info:
    get:
      operationId: getCalendarFeedInfo
      tags: [calendar]
      summary: Get the outbound calendar feed link
      description: Returns the private iCal subscription URL for authenticated calendar administrators.
      responses:
        "200":
          description: Outbound calendar feed link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarFeedInfo"
        "401":
          description: Admin sign-in required
        "403":
          description: Account is not an authorised admin
        "503":
          description: Outbound calendar feed is not configured
  /calendar/feed:
    get:
      operationId: getCalendarFeed
      tags: [calendar]
      summary: Read the merged outbound iCal feed
      description: Returns a token-protected iCal feed containing the merged blocked periods from the connected OTA calendars.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: exclude
          in: query
          required: false
          description: Omit events imported from this source to prevent feedback loops.
          schema:
            type: string
            enum: [bookingCom, airbnb, makeMyTrip]
      responses:
        "200":
          description: Merged iCal calendar
          content:
            text/calendar:
              schema:
                type: string
        "401":
          description: Invalid calendar feed token
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CalendarSyncInput:
      type: object
      properties:
        bookingCom:
          type: ["string", "null"]
        airbnb:
          type: ["string", "null"]
        makeMyTrip:
          type: ["string", "null"]
      additionalProperties: false
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
          enum: [manual, direct, bookingCom, airbnb, makeMyTrip]
        title:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        editable:
          type: boolean
      required: [id, source, title, note, startDate, endDate, editable]
    CalendarSourceStatus:
      type: object
      properties:
        source:
          type: string
          enum: [bookingCom, airbnb, makeMyTrip]
        label:
          type: string
        status:
          type: string
          enum: [connected, missing, error]
        eventCount:
          type: number
        message:
          type: string
        lastSyncedAt:
          type: ["string", "null"]
          format: date-time
      required: [source, label, status, eventCount, message, lastSyncedAt]
    CalendarSyncResponse:
      type: object
      properties:
        syncedAt:
          type: string
          format: date-time
        totalEvents:
          type: integer
        events:
          type: array
          items:
            $ref: "#/components/schemas/CalendarEvent"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/CalendarSourceStatus"
      required: [syncedAt, totalEvents, events, sources]
    CalendarFeedInfo:
      type: object
      properties:
        feedUrl:
          type: string
          format: uri
        bookingCom:
          type: string
          format: uri
        airbnb:
          type: string
          format: uri
        makeMyTrip:
          type: string
          format: uri
      required: [feedUrl, bookingCom, airbnb, makeMyTrip]
    CalendarBlockInput:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        title:
          type: ["string", "null"]
          maxLength: 120
        note:
          type: ["string", "null"]
          maxLength: 500
      required: [startDate, endDate]
    CalendarEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/CalendarEvent"
      required: [events]
    PublicCalendarBlock:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [startDate, endDate]
    PublicCalendarResponse:
      type: object
      properties:
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/PublicCalendarBlock"
      required: [blocks]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]