# lib/api-zod/src/generated/types/calendarBlockInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarBlockInput {
  startDate: Date;
  endDate: Date;
  /**
     * @maxLength 120
     * @nullable
     */
  title?: string | null;
  /**
     * @maxLength 500
     * @nullable
     */
  note?: string | null;
}