# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarEventSource } from './calendarEventSource';

export interface CalendarEvent {
  id: string;
  source: CalendarEventSource;
  /** @nullable */
  title: string | null;
  /** @nullable */
  note: string | null;
  startDate: Date;
  endDate: Date;
  editable: boolean;
}