# lib/api-zod/src/generated/types/calendarEventsResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarEvent } from './calendarEvent';

export interface CalendarEventsResponse {
  events: CalendarEvent[];
}