# lib/api-zod/src/generated/types/calendarSyncResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarEvent } from './calendarEvent';
import type { CalendarSourceStatus } from './calendarSourceStatus';

export interface CalendarSyncResponse {
  syncedAt: Date;
  totalEvents: number;
  events: CalendarEvent[];
  sources: CalendarSourceStatus[];
}