# lib/api-zod/src/generated/types/publicCalendarResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicCalendarBlock } from './publicCalendarBlock';

export interface PublicCalendarResponse {
  blocks: PublicCalendarBlock[];
}